requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
sentence-transformers
faiss-cpu
pydantic
tqdm